Added Java Classes: UserData and Characters
Characters class: this class is a TEMP-soluction and is incomplete and may change over time.
This class will extrend from a Creature Class.

UserData class: It contains all LOCAL Data Of the Currect User and The Data of the Characters
And it uses the constructor from the Character class to create new accounts.